refactor: Use common normalizeInput, move types to top level, remove msapplication-TileImage
