+++ src/exports/favicons.ts
@@ -5,5 +5,5 @@ export {
   googleS2,
 } from '../favicons/api/index.js'
 export { defaultGuessPaths, defaultIconRels } from '../favicons/defaults.js'
-export * from '../favicons/discover/types.js'
 export type { GuessMethodOptions } from '../favicons/guess/index.js'
+export * from '../favicons/types.js'
